cms/scripts/import-wins/pyproject.toml
[project]
name = "import-wins"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
